gpt_code_ui/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.4
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dateparser==1.1.8
debugpy==1.6.7.post1
decorator==5.1.1
docutils==0.20.1
et-xmlfile==1.1.0
executing==1.2.0
Fiona==1.9.4.post1
Flask==2.3.3
Flask-Cors==3.0.10
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==24.2.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mysql-connector-python==8.1.0
nest-asyncio==1.5.7
numpy==1.25.2
openai==0.28.0
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.18.0
pypng==0.20220715.0
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.1
qrcode==7.4.2
readme-renderer==41.0
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
shapely==2.0.1
six==1.16.0
snakeMQ==1.6
stack-data==0.6.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
SQLAlchemy==2.0.22
prettytable==3.9.0
pymysql==1.1.0
#vertica-python==1.3.6